Rename some services folders
