+++ ENGHelperBot/Services/Commands/Dictionaries/AddDictionaryCommand.cs
@@ -1,10 +1,11 @@
-﻿using ENGHelperBot.Services.Context;
+﻿using ENGHelperBot.Services.Command;
+using ENGHelperBot.Services.Context;
 using ENGHelperBot.Services.Repositories.Dictionaries;
 using Telegram.Bot;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 
-namespace ENGHelperBot.Services.Command;
+namespace ENGHelperBot.Services.Commands.Dictionaries;
 
 public class AddDictionaryCommand(IServiceScopeFactory scopeFactory) : ICommandHandler
 {
@@ -24,7 +25,7 @@ public class AddDictionaryCommand(IServiceScopeFactory scopeFactory) : ICommandH
 
         var chatContext = scope.ServiceProvider.GetRequiredService<IChatContextProvider>();
         chatContext.ResetFollowingCommand(update.Message!.Chat.Id);
-        
+
         var dictionaryService = scope.ServiceProvider.GetRequiredService<IDictionaryRepository>();
         var isSuccess = await dictionaryService.CreateAsync(new() { Name = update.Message!.Text!, UserId = update.Message.Chat.Id },
                                                             d => d.UserId == update.Message.Chat.Id);
