+++ ENGHelperBot/Services/Commands/ForwardCommand.cs
